Factorize stuff. Add Quiet mode. Check if the file exist.
