+++ sdk/driver/integration/suite_test.go
@@ -12,6 +12,7 @@ import (
 	"testing"
 	"time"
 
+	common "gopkg.in/bblfsh/sdk.v1"
 	"gopkg.in/bblfsh/sdk.v1/protocol"
 
 	"github.com/pmezard/go-difflib/difflib"
@@ -114,8 +115,10 @@ func (s *suite) doTestNativeParse(t *testing.T, filename string) {
 	r.Nil(err)
 
 	expected := getAST(r, filename)
+	resstr, err := common.NativeParseResponseToString(res)
+	r.Nil(err)
 
-	EqualText(r, expected, NativeParseResponseToString(r, res))
+	EqualText(r, expected, resstr)
 }
 
 func NativeParseResponseToString(r *require.Assertions, res *protocol.NativeParseResponse) string {
